// src/features/access-control/types/access-control.types.ts
export interface FullUserDto {
  id: number;
  firstName?: string;
  lastName?: string;
  fullName?: string;
  email?: string;
}
// ...
export interface PermissionGroupDto {
  id: number;
  createdDate: string;
  updatedDate?: string;
  deletedDate?: string;
  isDeleted: boolean;
  createdByFullUser?: FullUserDto;
  updatedByFullUser?: FullUserDto;
  deletedByFullUser?: FullUserDto;
  name: string;
  description?: string;
  isSystemAdmin: boolean;
  isActive: boolean;
  permissionDefinitionIds: number[];
  permissionCodes: string[];
  code?: string;
  priority?: number;
  isSystem?: boolean;
}